+++ lib/rails_pulse.rb
@@ -5,6 +5,7 @@ require "rails_pulse/configuration"
 require "rails_pulse/paginator"
 require "rails_pulse/cleanup_service"
 require "rails_pulse/tracker"
+require "rails_pulse/standalone"
 
 module RailsPulse
   class << self